refactor(): adjust to the repo coding guidelines
