+++ packages/microservices/utils/index.ts
@@ -1,1 +1,1 @@
-export * from './msvc.util';
+export * from './transform-pattern.utils';
